requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Django==4.0.3
django-archive==0.2.0
djangorestframework==3.13.1
flake8==4.0.1
idna==3.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
soupsieve==2.3.2
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.9
xmltodict==0.12.0
